Move two widgets functions over to proper source file
